requirements.txt
# Core
pandas==2.1.4
numpy==1.26.4

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Airflow (install separately via Docker or pip)
# apache-airflow==2.8.0

# Elasticsearch
elasticsearch==8.12.0

# Data Validation
great-expectations==0.18.12

# Visualization
matplotlib==3.8.2
seaborn==0.13.2

# Notebook
jupyter==1.0.0
ipykernel==6.29.2
